+++ src/backend/ci/core/auth/biz-auth-rbac/src/main/kotlin/com/tencent/devops/auth/service/RbacPermissionResourceGroupService.kt
@@ -29,8 +29,10 @@
 package com.tencent.devops.auth.service
 
 import com.tencent.bk.sdk.iam.constants.ManagerScopesEnum
+import com.tencent.bk.sdk.iam.dto.PageInfoDTO
 import com.tencent.bk.sdk.iam.dto.V2PageInfoDTO
 import com.tencent.bk.sdk.iam.dto.manager.ManagerRoleGroup
+import com.tencent.bk.sdk.iam.dto.manager.V2ManagerRoleGroupInfo
 import com.tencent.bk.sdk.iam.dto.manager.dto.GroupMemberRenewApplicationDTO
 import com.tencent.bk.sdk.iam.dto.manager.dto.SearchGroupDTO
 import com.tencent.bk.sdk.iam.service.v2.V2ManagerService
@@ -54,6 +56,8 @@ import com.tencent.devops.common.api.pojo.Pagination
 import com.tencent.devops.common.api.util.DateTimeUtil
 import com.tencent.devops.common.api.util.PageUtil
 import com.tencent.devops.common.auth.api.AuthResourceType
+import com.tencent.devops.common.auth.api.pojo.BkAuthGroup
+import com.tencent.devops.common.auth.api.pojo.BkAuthGroupAndUserList
 import com.tencent.devops.common.web.utils.I18nUtil
 import org.jooq.DSLContext
 import org.slf4j.LoggerFactory
@@ -318,6 +322,116 @@ class RbacPermissionResourceGroupService @Autowired constructor(
         iamV2ManagerService.updateRoleGroupV2(groupId, managerRoleGroup)
         return true
     }
+
+    override fun getResourceGroupUsers(
+        projectCode: String,
+        resourceType: String,
+        resourceCode: String,
+        group: String?
+    ): List<String> {
+        return when (group) {
+            // 新的rbac版本中，没有ci管理员组，不可以调用此接口来获取ci管理员组的成员
+            BkAuthGroup.CIADMIN.value, BkAuthGroup.CI_MANAGER.value -> emptyList()
+            // 获取特定资源下全部成员
+            null -> {
+                getResourceGroupAndUserList(
+                    projectCode = projectCode,
+                    resourceType = resourceType,
+                    resourceCode = resourceCode
+                ).flatMap { it.userIdList }.distinct()
+            }
+            else -> {
+                val dbGroupInfo = authResourceGroupDao.get(
+                    dslContext = dslContext,
+                    projectCode = projectCode,
+                    resourceType = resourceType,
+                    resourceCode = resourceCode,
+                    groupCode = group
+                ) ?: return emptyList()
+                val groupInfo = getResourceGroupAndUserList(
+                    projectCode = projectCode,
+                    resourceType = resourceType,
+                    resourceCode = resourceCode
+                ).find { it.roleId == dbGroupInfo.relationId.toInt() }
+                groupInfo?.userIdList ?: emptyList()
+            }
+        }
+    }
+
+    override fun getResourceGroupAndUserList(
+        projectCode: String,
+        resourceType: String,
+        resourceCode: String
+    ): List<BkAuthGroupAndUserList> {
+        // 1、获取管理员id
+        val managerId = authResourceService.get(
+            projectCode = projectCode,
+            resourceType = resourceType,
+            resourceCode = resourceCode
+        ).relationId
+        // 2、获取分级管理员下所有的用户组
+        val groupInfoList = getGroupInfoList(
+            resourceType = resourceType,
+            managerId = managerId
+        )
+        logger.info(
+            "[RBAC-IAM] getProjectGroupAndUserList: projectCode = $projectCode |" +
+                " gradeManagerId = $managerId | groupInfoList: $groupInfoList"
+        )
+        // 3、获取组成员
+        return groupInfoList.map { getUsersUnderGroup(groupInfo = it) }
+    }
+
+    private fun getGroupInfoList(
+        resourceType: String,
+        managerId: String
+    ): List<V2ManagerRoleGroupInfo> {
+        return if (resourceType == AuthResourceType.PROJECT.value) {
+            val searchGroupDTO = SearchGroupDTO.builder().inherit(false).build()
+            permissionGradeManagerService.listGroup(
+                gradeManagerId = managerId,
+                searchGroupDTO = searchGroupDTO,
+                page = 1,
+                pageSize = 1000
+            )
+        } else {
+            val v2PageInfoDTO = V2PageInfoDTO().apply {
+                pageSize = 1000
+                page = 1
+            }
+            iamV2ManagerService.getSubsetManagerRoleGroup(
+                managerId.toInt(),
+                v2PageInfoDTO
+            ).results
+        }
+    }
+
+    private fun getUsersUnderGroup(groupInfo: V2ManagerRoleGroupInfo): BkAuthGroupAndUserList {
+        val pageInfoDTO = PageInfoDTO().apply {
+            limit = 1000
+            offset = 0
+        }
+        val groupMemberInfoList = iamV2ManagerService.getRoleGroupMemberV2(groupInfo.id, pageInfoDTO).results
+        logger.info(
+            "[RBAC-IAM] getProjectGroupAndUserList ,groupId: ${groupInfo.id} " +
+                "| groupMemberInfoList: $groupMemberInfoList"
+        )
+        val members = mutableListOf<String>()
+        groupMemberInfoList.forEach { memberInfo ->
+            // todo 暂时不返回部门的用户
+            if (memberInfo.type == ManagerScopesEnum.getType(ManagerScopesEnum.USER)) {
+                members.add(memberInfo.id)
+            }
+        }
+        return BkAuthGroupAndUserList(
+            displayName = groupInfo.name,
+            roleId = groupInfo.id,
+            roleName = groupInfo.name,
+            userIdList = members.toSet().toList(),
+            type = ""
+        )
+    }
+
 
     private fun checkDuplicateGroupName(
         projectId: String,
