+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/resources/service/ServiceProjectAuthResourceImpl.kt
@@ -58,23 +58,6 @@ class ServiceProjectAuthResourceImpl @Autowired constructor(
         )
     }
 
-    override fun getResourceUsers(
-        token: String,
-        projectCode: String,
-        resourceType: String,
-        resourceCode: String,
-        group: BkAuthResourceGroup?
-    ): Result<List<String>> {
-        return Result(
-            permissionProjectService.getProjectUsers(
-                projectCode = projectCode,
-                resourceType = resourceType,
-                resourceCode = resourceCode,
-                group = group?.value
-            )
-        )
-    }
-
     override fun getProjectGroupAndUserList(
         token: String,
         projectCode: String
