+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/service/iam/PermissionResourceGroupService.kt
@@ -34,6 +34,7 @@ import com.tencent.devops.auth.pojo.vo.IamGroupInfoVo
 import com.tencent.devops.auth.pojo.vo.IamGroupMemberInfoVo
 import com.tencent.devops.auth.pojo.vo.IamGroupPoliciesVo
 import com.tencent.devops.common.api.pojo.Pagination
+import com.tencent.devops.common.auth.api.pojo.BkAuthGroupAndUserList
 
 interface PermissionResourceGroupService {
     /**
@@ -99,4 +100,17 @@ interface PermissionResourceGroupService {
         groupId: Int,
         renameGroupDTO: RenameGroupDTO
     ): Boolean
+
+    fun getResourceGroupUsers(
+        projectCode: String,
+        resourceType: String,
+        resourceCode: String,
+        group: String?
+    ): List<String>
+
+    fun getResourceGroupAndUserList(
+        projectCode: String,
+        resourceType: String,
+        resourceCode: String
+    ): List<BkAuthGroupAndUserList>
 }
